Add quotations
I added quotations. Now the '~' symbol can be used to prevent a substring of text from being translated into emojis. This is in addition to the backslash, which can be used to prevent the character immediately following it from being translated into emojis or being used as a special character. ~Safe emojing~ \.